--- src/inquirer/index.ts
@@ -1,3 +1,0 @@
-export * from './auth';
-export * from './pr';
-export * from './confirm';
